+++ R/tidy_add_header_rows.R
@@ -165,6 +165,7 @@ tidy_add_header_rows <- function(x,
 
     if (nrow(header_rows) > 0) {
       header_rows <- header_rows %>%
+        dplyr::mutate(term_cleaned = .clean_backticks(.data$term, .data$variable)) %>%
         dplyr::group_by(.data$variable, .data$y.level) %>%
         dplyr::summarise(
           var_class = dplyr::first(.data$var_class),
@@ -174,7 +175,7 @@ tidy_add_header_rows <- function(x,
           contrasts = dplyr::first(.data$contrasts),
           contrasts_type = dplyr::first(.data$contrasts_type),
           var_nrow = dplyr::n(),
-          var_test = sum(.data$term != .data$variable),
+          var_test = sum(.data$term_cleaned != .data$variable),
           rank = min(.data$rank) - .25,
           .groups = "drop_last"
         ) %>%
@@ -191,6 +192,7 @@ tidy_add_header_rows <- function(x,
 
     if (nrow(header_rows) > 0)
       header_rows <- header_rows %>%
+        dplyr::mutate(term_cleaned = .clean_backticks(.data$term, .data$variable)) %>%
         dplyr::group_by(.data$variable) %>%
         dplyr::summarise(
           var_class = dplyr::first(.data$var_class),
@@ -200,7 +202,7 @@ tidy_add_header_rows <- function(x,
           contrasts = dplyr::first(.data$contrasts),
           contrasts_type = dplyr::first(.data$contrasts_type),
           var_nrow = dplyr::n(),
-          var_test = sum(.data$term != .data$variable),
+          var_test = sum(.data$term_cleaned != .data$variable), # for dichotomous variables with no reference row
           rank = min(.data$rank) - .25,
           .groups = "drop_last"
         ) %>%